+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ### Fixed
 - Event query by isDelivered
 - Event async update
+### Added
+- workspace.update() to allow Workspace parameter updates
 
 ## [2.5.2] - 2021-04-23
 ### Fixed